[feature/DATEPICK-220]: BaseEntity 삭제 및 id 값들 모두 entity로 옮기기 + Entity 네이밍, 타입 수정

+++ src/main/java/app/hdj/datepick/domain/course/dto/CourseFilterParam.java
@@ -13,7 +13,7 @@ public class CourseFilterParam {
 
     private String keyword;
 
-    private List<@Positive Byte> tagId;
+    private List<@Positive Integer> tagId;
 
     @Positive
     private Long featuredId;
